requirements.txt
gunicorn==19.7.1
gevent==1.2.2
Django==2.0.2
git+https://github.com/jazzband/django-debug-toolbar.git#egg=django-debug-toolbar
django-filter==1.1.0
psycopg2-binary==2.7.4
Unipath==1.1
ipython==6.2.1
django-environ==0.4.4
dj-database-url==0.4.2
requests

# Tests
mixer==6.0.1
mock==2.0.0
